requirements.txt
mastodon.py
python-dotenv
beautifulsoup4
libvirt-python
opencv-python-headless
pytesseract